init curate processing

+++ image/systems/embark.py
@@ -2,7 +2,7 @@ import os
 import json
 from queue import Queue
 import threading
-from pyvips import Image
+from pyvips import Image, Error
 import botocore
 import shared.config as config
 import shared.logger as logger
@@ -46,43 +46,50 @@ def _reprocess_image(queue: Queue) -> None:
             global counter
             counter += 1
             image = _preprocess_image(img_data, local_file)
-            image.tiffsave(tif_filename, tile=True, pyramid=True, compression=config.COMPRESSION_TYPE,
-                tile_width=config.PYTIF_TILE_WIDTH, tile_height=config.PYTIF_TILE_HEIGHT, \
-                xres=config.DPI_VALUE, yres=config.DPI_VALUE) # noqa
+            if image:
+                image.tiffsave(tif_filename, tile=True, pyramid=True, compression=config.COMPRESSION_TYPE,
+                    tile_width=config.PYTIF_TILE_WIDTH, tile_height=config.PYTIF_TILE_HEIGHT, \
+                    xres=config.DPI_VALUE, yres=config.DPI_VALUE) # noqa
+                key = f"{path}/{tif_filename}"
+                aws_utility.upload_file(config.IMAGE_BUCKET, key, tif_filename)
+                os.remove(tif_filename)
             os.remove(local_file)
-            key = f"{path}/{tif_filename}"
-            aws_utility.upload_file(config.IMAGE_BUCKET, key, tif_filename)
-            os.remove(tif_filename)
         queue.task_done()
         # print(f"{image.get_fields()}")  # image fields, including exif
 
 
 def _preprocess_image(img_data: dict, local_file: str) -> Image:
-    image = Image.new_from_file(local_file, access='sequential')
-    max_height = config.DEFAULT_MAX_HEIGHT
-    max_width = config.DEFAULT_MAX_WIDTH
-    if img_data.get("copyrightStatus", "").lower() == "copyright":
-        max_height = config.COPYRIGHT_MAX_HEIGHT
-        max_width = config.COPYRIGHT_MAX_WIDTH
-    if image.height > max_height or image.width > max_width:
-        if image.height >= image.width:
-            shrink_by = image.height / max_height
-        else:
-            shrink_by = image.width / max_width
-        logger.debug(f"Image resizing - {img_data['id']}")
-        logger.debug(f"Resizing {local_file} by: {shrink_by}")
-        logger.debug(f"Original {local_file} height: {image.height}")
-        logger.debug(f"Original {local_file} width: {image.width}")
-        image = image.shrink(shrink_by, shrink_by)
-        logger.debug(f"Updated {local_file} height: {image.height}")
-        logger.debug(f"Updated {local_file} width: {image.width}")
+    try:
+        image = Image.new_from_file(local_file, access='sequential')
+        max_height = config.DEFAULT_MAX_HEIGHT
+        max_width = config.DEFAULT_MAX_WIDTH
+        if img_data.get("copyrightStatus", "").lower() == "copyright":
+            max_height = config.COPYRIGHT_MAX_HEIGHT
+            max_width = config.COPYRIGHT_MAX_WIDTH
+        if image.height > max_height or image.width > max_width:
+            if image.height >= image.width:
+                shrink_by = image.height / max_height
+            else:
+                shrink_by = image.width / max_width
+            logger.debug(f"Image resizing - {img_data['id']}")
+            logger.debug(f"Resizing {local_file} by: {shrink_by}")
+            logger.debug(f"Original {local_file} height: {image.height}")
+            logger.debug(f"Original {local_file} width: {image.width}")
+            image = image.shrink(shrink_by, shrink_by)
+            logger.debug(f"Updated {local_file} height: {image.height}")
+            logger.debug(f"Updated {local_file} width: {image.width}")
+    except Error as pye:
+        image = None
+        logger.error(f"VIPs error - {pye.message}")
     return image
 
 
 def process_embark_changes():
     jobs = Queue()
+    logger.info("GATHERING EMBARK IMAGES TO PROCESS")
     for img_data in _list_changes():
         jobs.put(img_data)
+    logger.info(f"{jobs.qsize()} IMAGES TO PROCESS")
 
     start_time = time.time()
     for i in range(config.MAX_THREADS):
